pyproject.toml
[project]
name = "hu-ai-platform-demo"
version = "0.1.0"
description = "AI Platform Demo for Hogeschool Utrecht — RAG chatbot, agents, guardrails, and monitoring"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Hilario Pedro" }]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "agno>=1.0",
    "openai>=1.60",
    "pydantic-settings>=2.7",
    "faiss-cpu>=1.9",
    "tiktoken>=0.8",
    "opentelemetry-api>=1.29",
    "opentelemetry-sdk>=1.29",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "python-multipart>=0.0.18",
    "httpx>=0.28",
    "aiosqlite>=0.21",
    "numpy>=1.26",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "mypy>=1.14",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hu_ai_platform"]

[tool.ruff]
line-length = 99
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.ruff.lint.per-file-ignores]
"src/hu_ai_platform/api/**" = ["B008"]  # Depends() in defaults is idiomatic FastAPI

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

